move methods to units module
