+++ opgee/core.py
@@ -12,35 +12,11 @@ from typing import Optional, TypeVar
 
 import pint
 
+from opgee.units import validate_unit
+
 from .error import AbstractMethodError, ModelValidationError, OpgeeException
-from .log import getLogger
 from .units import ureg
 from .utils import coercible, getBooleanXML
-
-_logger = getLogger(__name__)
-
-def magnitude(value, units=None):
-    """
-    Return the magnitude of ``value``. If ``value`` is a ``pint.Quantity`` and
-    ``units`` is not None, check that ``value`` has the expected units and
-    return the magnitude of ``value``. If ``value`` is not a ``pint.Quantity``,
-    just return it.
-
-    :param value: (float or pint.Quantity) the value for which we return the magnitude.
-    :param units: (None or pint.Unit) the expected units
-    :return: the magnitude of `value`
-    """
-    if isinstance(value, ureg.Quantity):
-        # if optional units are provided, validate them
-        if units:
-            if not isinstance(units, pint.Unit):
-                units = ureg.Unit(units)
-            if value.units != units:
-                raise OpgeeException(f"magnitude: value {value} units are not {units}")
-
-        return value.m
-    else:
-        return value
 
 
 def name_of(obj):
@@ -221,32 +197,6 @@ class XmlInstantiable(OpgeeObject):
             return None
 
         return self.parent.find_container(cls)  # recursively ascend the graph
-
-
-# to avoid redundantly reporting bad units
-_undefined_units = {}
-
-
-def validate_unit(unit):
-    """
-    Return the ``pint.Unit`` associated with the string ``unit``, or ``None``
-    if ``unit`` is ``None`` or not in the unit registry.
-
-    :param unit: (str) a string representation of a ``pint.Unit``
-
-    :return: (pint.Unit or None)
-    """
-    if not unit:
-        return None
-
-    if unit in ureg:
-        return ureg.Unit(unit)
-
-    if unit not in _undefined_units:
-        _logger.warning(f"Unit '{unit}' is not in the UnitRegistry")
-        _undefined_units[unit] = 1
-
-    return None
 
 
 class A(OpgeeObject):
